feat(#975): Extract JobCard into separate component
Extracted JobCard component from jobs page into its own file for better reusability and maintainability. Removed unnecessary constants that are now scoped to the component.
